[XFS] Remove some off_t abuse in pagebuf_offset and the page_io routine, after some careful analysis.
SGI Modid: 2.5.x-xfs:slinx:140039a

+++ fs/xfs/pagebuf/page_buf.h
@@ -329,7 +329,7 @@ extern int pagebuf_iorequest(		/* start real I/O		*/
 extern int pagebuf_iowait(		/* wait for buffer I/O done	*/
 		page_buf_t *);		/* buffer to wait on		*/
 
-extern caddr_t	pagebuf_offset(page_buf_t *, off_t);
+extern caddr_t	pagebuf_offset(page_buf_t *, size_t);
 
 extern void pagebuf_iomove(		/* move data in/out of pagebuf	*/
 		page_buf_t *,		/* buffer to manipulate		*/
